Rename `appng` package to `app`
