+++ v2/internal/app/app_default_darwin.go
@@ -1,7 +1,7 @@
-//go:build !dev && !production && !bindings && linux
-// +build !dev,!production,!bindings,linux
+//go:build !dev && !production && !bindings && darwin
+// +build !dev,!production,!bindings,darwin
 
-package appng
+package app
 
 import (
 	"fmt"
